+++ crates/holochain/src/core/workflow/initialize_zomes_workflow.rs
@@ -71,7 +71,11 @@ where
     // Insert the init marker
     workspace
         .source_chain()
-        .put(builder::InitZomesComplete {}, None)
+        .put(
+            builder::InitZomesComplete {},
+            None,
+            ChainTopOrdering::Strict,
+        )
         .await?;
 
     // TODO: Validate scratch items
@@ -220,13 +224,14 @@ pub mod tests {
     #[tokio::test(flavor = "multi_thread")]
     async fn commit_during_init_one_zome_unimplemented_one_fails() {
         let zome_fail = InlineZome::new_unique(vec![]).callback("init", |api, _: ()| {
-            api.create(EntryWithDefId::new(
+            api.create(CreateInput::new(
                 EntryDefId::CapGrant,
                 Entry::CapGrant(CapGrantEntry {
                     tag: "".into(),
                     access: ().into(),
                     functions: vec![("no-init".into(), "xxx".into())].into_iter().collect(),
                 }),
+                ChainTopOrdering::default(),
             ))?;
             Ok(InitCallbackResult::Fail("reason".into()))
         });