+++ crates/holochain/src/sweettest/sweet_cell.rs
@@ -1,13 +1,16 @@
 use super::SweetZome;
 use hdk::prelude::*;
 use holo_hash::DnaHash;
-use holochain_types::env::EnvWrite;
+use holochain_p2p::{dht_arc::ArcInterval, AgentPubKeyExt, HolochainP2pCell};
+use holochain_types::prelude::*;
+use kitsune_p2p::actor::TestBackdoor;
 /// A reference to a Cell created by a SweetConductor installation function.
 /// It has very concise methods for calling a zome on this cell
 #[derive(Clone, derive_more::Constructor)]
 pub struct SweetCell {
     pub(super) cell_id: CellId,
     pub(super) cell_env: EnvWrite,
+    pub(super) network: HolochainP2pCell,
 }
 
 impl SweetCell {
@@ -34,5 +37,21 @@ impl SweetCell {
     /// Get a SweetZome with the given name
     pub fn zome<Z: Into<ZomeName>>(&self, zome_name: Z) -> SweetZome {
         SweetZome::new(self.cell_id.clone(), zome_name.into())
+    }
+
+    /// Coerce the agent's storage arc to the specified value.
+    /// The arc need not be centered on the agent's DHT location, which is
+    /// typically a requirement "in the real world", but this can be useful
+    /// for integration tests of gossip.
+    ///
+    #[cfg(feature = "test_utils")]
+    pub async fn set_storage_arc(&self, arc: ArcInterval) {
+        use holochain_p2p::HolochainP2pCellT;
+
+        let agent = self.cell_id.agent_pubkey().to_kitsune();
+        self.network
+            .test_backdoor(TestBackdoor::SetArc(agent, arc))
+            .await
+            .unwrap();
     }
 }