+++ crates/kitsune_p2p/direct/src/handle_ws.rs
@@ -234,6 +234,29 @@ impl AsKdHnd for Hnd {
         .boxed()
     }
 
+    fn is_authority(
+        &self,
+        root: KdHash,
+        agent: KdHash,
+        basis: KdHash,
+    ) -> BoxFuture<'static, KdResult<bool>> {
+        let msg_id = new_msg_id();
+        let api = KdApi::IsAuthorityReq {
+            msg_id,
+            root,
+            agent,
+            basis,
+        };
+        let api = self.request(api);
+        async move {
+            match api.await {
+                Ok(KdApi::IsAuthorityRes { is_authority, .. }) => Ok(is_authority),
+                oth => Err(format!("unexpected: {:?}", oth).into()),
+            }
+        }
+        .boxed()
+    }
+
     fn message_send(
         &self,
         root: KdHash,